@@ -379,11 +379,17 @@ public sealed class DiscordRichPresence
         {
             if (lid.StartsWith("spotify:"))
             {
-                this.LargeImage = new DiscordSpotifyAsset { Id = lid };
+                this.LargeImage = new DiscordSpotifyAsset(lid);
             }
             else if (ulong.TryParse(lid, NumberStyles.Number, CultureInfo.InvariantCulture, out ulong ulid))
             {
-                this.LargeImage = new DiscordApplicationAsset { Id = lid, Application = this.Application, Type = ApplicationAssetType.LargeImage };
+                this.LargeImage =
+                    new DiscordApplicationAsset
+                    {
+                        Id = lid,
+                        Application = this.Application,
+                        Type = DiscordApplicationAssetType.LargeImage
+                    };
             }
         }
 
@@ -392,11 +398,17 @@ public sealed class DiscordRichPresence
         {
             if (sid.StartsWith("spotify:"))
             {
-                this.SmallImage = new DiscordSpotifyAsset { Id = sid };
+                this.SmallImage = new DiscordSpotifyAsset(sid);
             }
             else if (ulong.TryParse(sid, NumberStyles.Number, CultureInfo.InvariantCulture, out ulong usid))
             {
-                this.SmallImage = new DiscordApplicationAsset { Id = sid, Application = this.Application, Type = ApplicationAssetType.LargeImage };
+                this.SmallImage =
+                    new DiscordApplicationAsset
+                    {
+                        Id = sid,
+                        Application = this.Application,
+                        Type = DiscordApplicationAssetType.SmallImage
+                    };
             }
         }
     }
